+++ app/src/main/java/com/davidmag/movienatic/domain/repository/MovieRepository.kt
@@ -6,7 +6,7 @@ import io.reactivex.Maybe
 
 interface MovieRepository {
     fun search(query : String) : Maybe<List<Movie>>
-    fun fetch(genreId : Int? = null) : Maybe<Any>
-    fun find(id : Int) : Maybe<Any>
-    fun get(id : Int? = null, genreId : Int? = null) : Flowable<List<Movie>>
+    fun fetch(genreId : Long? = null) : Maybe<Any>
+    fun find(id : Long) : Maybe<Any>
+    fun get(id : Long? = null, genreId : Long? = null) : Flowable<List<Movie>>
 }
