+++ app/src/main/java/com/davidmag/movienatic/domain/usecase/FetchMovieDetailsByIdUseCase.kt
@@ -6,7 +6,7 @@ import io.reactivex.Maybe
 class FetchMovieDetailsByIdUseCase (
     val movieRepository: MovieRepository
 ) {
-    fun execute(id : Int) : Maybe<*> {
+    fun execute(id : Long) : Maybe<*> {
         return movieRepository.find(id)
     }
 }
