+++ app/src/main/java/com/davidmag/movienatic/domain/usecase/GetMoviesUseCase.kt
@@ -7,7 +7,7 @@ import io.reactivex.Flowable
 class GetMoviesUseCase(
     val movieRepository: MovieRepository
 ) {
-    fun execute(id : Int? = null, genreId : Int? = null) : Flowable<List<Movie>> {
+    fun execute(id : Long? = null, genreId : Long? = null) : Flowable<List<Movie>> {
         return movieRepository.get(id, genreId)
     }
 }
